refactor(messaging): remove conflict adapter
